LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/prefetch-status.https_status=400&should_prefetch=false-expected.txt


FAIL Check that only prefetched requests with status in 200-299 range are used. assert_equals: Prefetch should not work for request statue:400. expected (undefined) undefined but got (string) "prefetch"
